facelifted register and login page. Fixed route issues with both. Moved script.js to its own folder in FrontEnd
